Move utils.js into src/lib, rm unused functions, and break it into separate modules. - Finally organize utils.js.
