+++ src/pages/Index.vue
@@ -65,15 +65,8 @@
 import HomeTop from "@/components/HomeTop";
 import HomeCard from "@/components/HomeCard";
 import HomeProfile from "@/components/HomeProfile.vue";
-import {
-    hasContent,
-    gatherCollections,
-    gatherInserts,
-    gatherCards,
-    makeCardRows,
-    addTwitterWidget,
-    addAltmetrics,
-} from "~/utils.js";
+import { hasContent, gatherCollections, gatherInserts, gatherCards, makeCardRows } from "~/lib/pages.mjs";
+import { addTwitterWidget, addAltmetrics } from "~/lib/client.mjs";
 export default {
     components: {
         HomeTop,
